feat: add Profile CRD to this charm, update tests and status logging
This commit moves the profile CRD into this charm (previously it was in the kubeflow-dashboard charm) as well as other misc fixes.

History: The Profile CRD has previously been deployed by the kubeflow-dashboard charm due to their interdependency:
* kubeflow-dashboard expects to have a Profile backend to communicate with, and historically needed to create a Profile object for the admin user

In the current state of Kubeflow, however, dashboard does not require a profile to be generated (it will generate them on the fly) and the profiles-operator may be desirable to run standalone without dashboards (eg: if we run just some backend services that need profiles maintained, or if we upgrade dashboards and don't want to delete all profile objects).  Given this, it makes sense to move the Profile object to the controller that manages it.

Other changes:
* Update Profile CRD to v1
* Update integration tests to test if profile-controller is generating namespaces for profiles, and to clean up after the test to avoid test leakage
* pin microk8s to 1.21 in integrate.yaml (upstream does not support >1.21)
* fix status reporting to avoid __init__ overwriting existing error Status' (see below).
* fix race condition(?) in relations by bumping ops to 1.2.0.  Accessing relation data from the relation data bag was (we think?) blocked by a race condition in the ops library.  Bumping to ops==1.2.0 fixes the issue

Status reporting:
Setting ActiveStatus() in init caused the accidental resetting of still valid statuses. For example:
* `config_change` hook triggers and is handled by `main()`.  Say it then hits an `OCIImageResourceError` in `main()`, setting an `ErrorStatus()`
* `update_status` hook triggers and has no handler.  It still runs `__init__()` which passes the `except NoVersionsListed/except NoCompatibleVersions` block, which sets `ActiveStatus`.  There is no additional handling on `update_status`, so it ends with the app in `ActiveStatus`

This masks any errors triggered by other hooks, making it harder for CI and humans to notice and fix errors (this also might happen instantly if an `update_status` is queued behind some other hook.