build: generate requirements-unit.txt with uv and pyproject.toml
This commit refactors the way the requirements-unit.txt file is generated by
removing its dependency with requirements.txt (now deleted because we are using pyproject.toml file)
and adding a dependency on the pyproject.toml file.
This is a support change for the usage of the uv plugin to build the charm.